+++ BlogposterCMS/jest.config.js
@@ -1,7 +1,18 @@
 module.exports = {
   testEnvironment: 'node',
   transform: {
-    '^.+\\.[tj]sx?$': 'babel-jest',
+    '^.+\\.[tj]sx?$': ['ts-jest', {
+      diagnostics: false,
+      tsconfig: {
+        allowJs: true,
+        target: 'ES2020',
+        module: 'commonjs',
+        moduleResolution: 'node',
+        esModuleInterop: true,
+        skipLibCheck: true,
+      },
+    }],
   },
+  moduleFileExtensions: ['ts', 'tsx', 'js', 'jsx', 'json', 'node'],
   transformIgnorePatterns: ['/node_modules/'],
 };